refactor(core): move createGameReducer into dedicated reducer.ts
Move createGameReducer and CreateGameReducerOptions out of types.ts into
a new core/src/reducer.ts module. Public API is unchanged (both are still
re-exported from @couch-kit/core) and runtime behaviour is identical.
Also removes a circular import between types.ts and middleware.ts.

Closes #45
