+++ packages/core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
+export * from "./reducer";
 export * from "./protocol";
 export * from "./constants";
 export * from "./replay";
